Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 int main() {
     int a,b,c;
     int sum=a+b+c;
-    int Average=sum/3;
+    float Average=sum/3.0;
     scanf("%d",&a,&b,&c);
     printf("Average: %.2d",Average);
     return 0;
